Update Restaurant Event Commit
Continued to updated the ResturantEvent Script, so it now allows for the human character to participate in the event, as well as a win and defeat conditions. The following script just needs to communicate with eachother; EventResult, PpointManager, and EventMonitor Scripts. When this is done, the Restaurant Event will be complete.